Refactor Session and Connector
- Makes Session the main API and Connector an implementation detail
- Connector is now a trait: other connectors can be swapped in
- The old connector struct is now ReqwestConnector
- ReqwestConnector is gated as a cargo feature
