+++ Cargo.toml
@@ -28,7 +28,7 @@ failure_derive = "0.1"
 hmac = "0.4"
 pbkdf2 = "0.1"
 rand = "0.4"
-reqwest = "0.8"
+reqwest = { version = "0.8", optional = true }
 serde = "1.0"
 serde_derive = "1.0"
 sha2 = "0.6"
@@ -36,9 +36,10 @@ tiny_http = { version = "0.5", optional = true }
 
 [features]
 bench = []
-default = []
+default = ["reqwest-connector"]
 dalek = ["ed25519-dalek"]
 mockhsm = ["dalek", "tiny_http"]
+reqwest-connector = ["reqwest"]
 
 [package.metadata.docs.rs]
-rustc-args = ["-C", "target-feature=+aes"]
+rustc-args = ["-Ctarget-feature=+aes"]
